requirements.txt
flask==2.3.3
flask-cors==4.0.0
supabase==1.3.0
pandas==2.1.3
plotly==5.17.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
wtforms==3.1.0
flask-wtf==1.2.1
reportlab==4.0.7
weasyprint==60.2
